Added confirmation logic to rm updated draw and remove
Im a little bit unsure on what the resposne List was meant ot be used for (created by me to long ago) so I don't think I have ended up ever using it and instead used the readyToSendRequest for more than I probably intended to use it for, However I believe this should work. could use some verification that the reasoning is correct.